# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Competitor Insight Hub API",
    "description": "API для мониторинга новостей из мира ИИ-индустрии",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/news/": {
      "post": {
        "tags": [
          "News",
          "news"
        ],
        "summary": "Create News",
        "operationId": "create_news_api_v1_news__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsCreateSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create News Api V1 News  Post"
                }
              }
            }
          },
          "404": {
            "description": "News item not found"
          },
          "422": {
            "description": "Invalid news parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "News",
          "news"
        ],
        "summary": "Get News",
        "description": "Get news items with enhanced filtering and search capabilities\n\nReturns paginated list of news items with comprehensive filtering options.",
        "operationId": "get_news_api_v1_news__get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NewsCategory"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by news category",
              "title": "Category"
            },
            "description": "Filter by news category"
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by single company ID",
              "title": "Company Id"
            },
            "description": "Filter by single company ID"
          },
          {
            "name": "company_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by multiple company IDs (comma-separated)",
              "title": "Company Ids"
            },
            "description": "Filter by multiple company IDs (comma-separated)"
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SourceType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source type",
              "title": "Source Type"
            },
            "description": "Filter by source type"
          },
          {
            "name": "search_query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search query for title/content",
              "title": "Search Query"
            },
            "description": "Search query for title/content"
          },
          {
            "name": "min_priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum priority score",
              "title": "Min Priority"
            },
            "description": "Minimum priority score"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of news items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of news items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of news items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of news items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get News Api V1 News  Get"
                }
              }
            }
          },
          "404": {
            "description": "News item not found"
          },
          "422": {
            "description": "Invalid news parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/news/{news_id}": {
      "put": {
        "tags": [
          "News",
          "news"
        ],
        "summary": "Update News",
        "operationId": "update_news_api_v1_news__news_id__put",
        "parameters": [
          {
            "name": "news_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "News Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update News Api V1 News  News Id  Put"
                }
              }
            }
          },
          "404": {
            "description": "News item not found"
          },
          "422": {
            "description": "Invalid news parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "News",
          "news"
        ],
        "summary": "Delete News",
        "operationId": "delete_news_api_v1_news__news_id__delete",
        "parameters": [
          {
            "name": "news_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "News Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "News item not found"
          },
          "422": {
            "description": "Invalid news parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "News",
          "news"
        ],
        "summary": "Get News Item",
        "description": "Get specific news item by ID with full details\n\nReturns detailed information about a specific news item including\nrelated company information, keywords, and user activities.",
        "operationId": "get_news_item_api_v1_news__news_id__get",
        "parameters": [
          {
            "name": "news_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "News Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get News Item Api V1 News  News Id  Get"
                }
              }
            }
          },
          "404": {
            "description": "News item not found"
          },
          "422": {
            "description": "Invalid news parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/news/stats": {
      "get": {
        "tags": [
          "News",
          "news"
        ],
        "summary": "Get News Statistics",
        "description": "Get comprehensive news statistics\n\nReturns statistics about news items including counts by category,\nsource type, recent news, and high priority items.",
        "operationId": "get_news_statistics_api_v1_news_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsStatsSchema"
                }
              }
            }
          },
          "404": {
            "description": "News item not found"
          },
          "422": {
            "description": "Invalid news parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/news/categories/list": {
      "get": {
        "tags": [
          "News",
          "news"
        ],
        "summary": "Get News Categories",
        "description": "Get available news categories with descriptions\n\nReturns list of all available news categories with their descriptions.",
        "operationId": "get_news_categories_api_v1_news_categories_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "News item not found"
          },
          "422": {
            "description": "Invalid news parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NewsCategory": {
        "type": "string",
        "enum": [
          "product_update",
          "pricing_change",
          "strategic_announcement",
          "technical_update",
          "funding_news",
          "research_paper",
          "community_event",
          "partnership",
          "acquisition",
          "integration",
          "security_update",
          "api_update",
          "model_release",
          "performance_improvement",
          "feature_deprecation"
        ],
        "title": "NewsCategory",
        "description": "Enhanced News category enumeration with descriptions"
      },
...
      "SourceType": {
        "type": "string",
        "enum": [
          "blog",
          "twitter",
          "github",
          "reddit",
          "news_site",
          "press_release"
        ],
        "title": "SourceType",
        "description": "Enhanced Source type enumeration with descriptions"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}